Introduce deploy parameters and initial .deb file support
In context of #1

An optional 'parameters' keyword argument has been added to deploy().
Pass a dictionary into it to attach parameters to deployed file.

Usage:
  p.deploy(stream, parameters={'a': 'b', 'c': ['d', 'e']})

Both strings and lists of strings are supported. Keys and values with
'=', ';' or ',' in their names are not yet supported.

Deploying .deb files can be done with with deploy_deb(), which is just
a thin wrapper around deploy() with parameters. Use it like this:

p.deploy_deb("./libatk1.0_i386.deb",
             distribution='ubuntu',
             component=['main', 'contrib'],
             architecture='i386')

You can specify lists and strings in distribution, component and
architecture.