Support for custom base URL
For uses who don't have /artifactory/ in base URL,
the library has failed to split the path properly,
because it uses /artifactory/ as heuristic for where
the 'root' ends and 'drive' begins.

As I didn't want to over-complicate the interface
of ArtifactoryPath (if it's even possible due to
pathlib constraints), I've implemented support for
custom roots via the configuration.

To tell the library your custom base URL, just put
that URL to ~/.artifactory_python.cfg like this:

[http://custom/base]

[https://another/custom/base]
verify = false

As you can see, other options can be specified as usual.
But then when you do ArtifactoryPath('custom/base/foo'),
it will be parsed as expected.

Fixes #22