Specifically handle JSONDecodeError as returned by requests
requests has a compat module that will use simplejson if it is available
within the environment. We should catch the exception outlined in this
module as it will reflect the json decoder used by requests.

Note that requests.exceptions.JSONDecodeError, and subsequently
requests.JSONDecodeError is not the same thing, and will not represent
the same exception in this case.