refactor: change ArtifactoryFlavour to behave like UnixFlavour
When checking the root in a UnixFlavour system, it is / for an absolute
path, and '' for a relative one, and unix systems have no drive.

For artifactory, if we consider the drive to be "http://b/artifactory"
and root to be "/", it behaves exactly like Windows and Unix Flavours,
and no if's and exceptions to handle the root are needed in the code.

This refactor makes the migration to python 3.12 much easier, since we
can subclass pathlib.Path directly and we don't need to override as many
methods.

Notice: This is a breaking change if you are accessing the `root` property
directly, but I'm assuming that mostly everyone should be using the
`repo` accessor for that.
There is also some minor changes in leading/trailing slashes