Reorganize code based on python 3.11 and 3.12
This commit addresses the concerns raised in the Pull request. It helps minimize
the amount of changes (way less indentation) compared to master.

Also, used a single constant for marking code that will be removed once 3.11 is
deprecated.

This commit likely should be "merged" into the previous two once the PR review
process is over. And if not, well, you have this nice comment here forever :)