Python 3.12 compat (#463)
* refactor: change ArtifactoryFlavour to behave like UnixFlavour

When checking the root in a UnixFlavour system, it is / for an absolute
path, and '' for a relative one, and unix systems have no drive.

For artifactory, if we consider the drive to be "http://b/artifactory"
and root to be "/", it behaves exactly like Windows and Unix Flavours,
and no if's and exceptions to handle the root are needed in the code.

This refactor makes the migration to python 3.12 much easier, since we
can subclass pathlib.Path directly and we don't need to override as many
methods.

Notice: This is a breaking change if you are accessing the `root` property
directly, but I'm assuming that mostly everyone should be using the
`repo` accessor for that.
There is also some minor changes in leading/trailing slashes

* feat: Code works under python3.12, but not under older versions of
python

* fix: Make project work in 3.11 and 3.12

Also, if tox is installed with python3.12, some pre-commit plugins
didn't work, so I have updated those.

* Reorganize code based on python 3.11 and 3.12

This commit addresses the concerns raised in the Pull request. It helps minimize
the amount of changes (way less indentation) compared to master.

Also, used a single constant for marking code that will be removed once 3.11 is
deprecated.

This commit likely should be "merged" into the previous two once the PR review
process is over. And if not, well, you have this nice comment here forever :)

---------

Co-authored-by: allburov <allburov@gmail.com>